feat(core): add level as an adapter option (#117)
* feat(core): add level as an adapter option

* chore(core/adapters): add better annotations to base adapter

* fix(core): account for iscale if disabling collect, add level tests

* chore(core/Divider): use series struct mem, account for iscale in combine if collect has been disabled

* fix(core/adapters): distribute rewards regardless of curr value

* chore(ci): bump forge version

* fix(ci): actually bump forge version

* refactor: begin using constructor-based adapters

* refactor: use constructor-based adapters

* refactor: re-add getStakeAndTarget, remove commented code and fix tests

* refactor(core/Divider): split level declaration out into a lib

* refactor(core/Divider): set level as adapter state var

* chore: mv levels lib to utils, add optional zero redeem hook

* chore: appease linter, add optimizer to all test cmds

* refactor(core/adapters): pack immutables

* fix(justfile): use flag syntax for optimize runs

* refactor(core/Divider): use levels for caller restrictions rather than enable/disable

* refactor(core/Divider): add tests for call restricted levels

* chore(core/Divider): use conditional rather than revert, appease linter

* fix(core/Divider): assign to struct correctly

* refactor(core/Divider): dont assign local var struct unnecessarily

* fix(core/Divider): resolve icap merge conflicts

+++ pkg/core/src/tests/BaseFactory.t.sol
@@ -30,14 +30,15 @@ contract Factories is TestHelper {
         assertEq(MockFactory(someFactory).divider(), address(divider));
         (
             address oracle,
-            uint256 ifee,
+            uint64 ifee,
             address stake,
             uint256 stakeSize,
-            uint256 minm,
-            uint256 maxm,
-            uint8 mode,
-            uint128 tilt
+            uint48 minm,
+            uint48 maxm,
+            uint16 mode,
+            uint64 tilt
         ) = MockFactory(someFactory).factoryParams();
+
         assertEq(oracle, ORACLE);
         assertEq(stake, address(stake));
         assertEq(ifee, ISSUANCE_FEE);
